Use doubles instead of floats, cleanup

+++ source/color.c
@@ -42,7 +42,7 @@ static void	color_init(t_color *c, uint32_t color)
 	c->red = (color >> 24 & 0xFF);
 }
 
-uint32_t	color_get(uint32_t start, uint32_t end, float percent)
+uint32_t	color_get(uint32_t start, uint32_t end, double percent)
 {
 	uint32_t	rgb[3];
 	t_color s;
@@ -50,8 +50,8 @@ uint32_t	color_get(uint32_t start, uint32_t end, float percent)
 
 	color_init(&s, start);
 	color_init(&e, end);
-	rgb[0] = (uint32_t)roundf((1 - percent) * s.red + percent * e.red);
-	rgb[1] = (uint32_t)roundf((1 - percent) * s.green + percent * e.green);
-	rgb[2] = (uint32_t)roundf((1 - percent) * s.blue + percent * e.blue);
+	rgb[0] = (uint32_t)round((1 - percent) * s.red + percent * e.red);
+	rgb[1] = (uint32_t)round((1 - percent) * s.green + percent * e.green);
+	rgb[2] = (uint32_t)round((1 - percent) * s.blue + percent * e.blue);
 	return 0xFF | (rgb[2] << 8) | ( rgb[1] << 16) | (rgb[0] << 24);
 }
